+++ Src/C/ipc/ewb/ewb_transfer.c
@@ -72,7 +72,7 @@ rt_public void ewb_tpipe(STREAM *stream) {
 #endif
 }
 
-rt_public char *ewb_tread(int *size)
+rt_public char *ewb_tread(size_t *size)
           		/* Filled in with size of read string */
 {
 	/* Read bytes from the "pipe" and put them into a new allocated buffer.
